+++ src/find.ts
@@ -10,39 +10,39 @@ import { Color, Text } from './interfaces';
 
 export { assertNo };
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function alert(text: Text, color?: Color) {
-  return cy.gears(components.Alert, text);
+  return cy.component(components.Alert, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function blockPanel(text: Text) {
-  return cy.gears(components.BlockPanel, text);
+  return cy.component(components.BlockPanel, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function button(text: Text) {
-  return cy.gears(components.Button, text);
+  return cy.component(components.Button, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function card(text: Text) {
-  return cy.gears(components.Card, text);
+  return cy.component(components.Card, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function cardTitle(text: Text) {
-  return cy.gears(components.Card, text).find('.card-title');
+  return cy.component(components.Card, text).find('.card-title');
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function checkboxInput(text: Text) {
-  return cy.gears(components.Input, text);
+  return cy.component(components.Input, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function datapair(text: Text) {
-  return cy.gears(components.Datapair, text);
+  return cy.component(components.Datapair, text);
 }
 
 export function hasManyFields(text: Text) {
@@ -51,38 +51,38 @@ export function hasManyFields(text: Text) {
   return cy.contains('label', text).parent();
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function input(text: Text) {
-  return cy.gears(components.Input, text);
+  return cy.component(components.Input, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function link(text: Text) {
-  return cy.gears(components.Link, text);
+  return cy.component(components.Link, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function modal(text: Text) {
-  return cy.gears(components.Modal, text);
+  return cy.component(components.Modal, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function modalTitle(text: Text) {
-  return cy.gears(components.Modal, text).find('.modal-title');
+  return cy.component(components.Modal, text).find('.modal-title');
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function radioInput(text: Text) {
   // TODO: type option
-  return cy.gears(components.Input, text);
+  return cy.component(components.Input, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function select(text: Text) {
-  return cy.gears(components.Select, text);
+  return cy.component(components.Select, text);
 }
 
-// @deprecated please use cy.gears() with Component definitions
+// @deprecated please use cy.component() with Component definitions
 export function summaryBoxItem(text: Text) {
-  return cy.gears(components.SummaryBoxItem, text);
+  return cy.component(components.SummaryBoxItem, text);
 }